requirements.txt
python-dotenv
pymongo
requests
Pillow